pyproject.toml
[project]
name = "datathonapp"
version = "0.1.0"
description = "Streamlit App para análise de dados e previsão das PEDEs da Assoc. Passos Mágicos"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "autogluon>=1.1.1",
    "keras>=3.5.0",
    "matplotlib>=3.9.2",
    "plotly>=5.24.1",
    "seaborn>=0.13.2",
    "streamlit-aggrid>=1.0.5",
    "streamlit-pandas-profiling>=0.1.3",
    "streamlit>=1.38.0",
]

[tool.uv]
dev-dependencies = [
    "bokeh>=3.5.2",
    "ipykernel>=6.29.5",
    "ipython>=8.27.0",
    "ipywidgets>=8.1.5",
    "nbformat>=5.10.4",
    "watchdog>=4.0.2",
]
[tool.pyright]
venvPath = "."
venv = ".venv"
